Protege os endpoints GET, PUT e DELETE com autenticação

+++ tests/test_app.py
@@ -40,32 +40,26 @@ def test_create_user_username_already_taken(client, user):
     assert response.json() == {'detail': 'Username already exists'}
 
 
-def test_create_user_email_already_take(client, user):
+def test_create_user_email_already_taken(client, user):
     response = client.post('/users/', json={'username': 'username', 'email': user.email, 'password': 'password'})
 
     assert response.status_code == HTTPStatus.CONFLICT
     assert response.json() == {'detail': 'Email already exists'}
 
 
-def test_read_users(client):
-    response = client.get('/users/')
-
-    assert response.status_code == HTTPStatus.OK
-    assert response.json() == {'users': []}
-
-
-def test_read_users_with_users(client, user):
+def test_read_users(client, user, token):
     user_schema = UserPublic.model_validate(user).model_dump()
-    response = client.get('/users/')
+    response = client.get('/users/', headers={'Authorization': f'Bearer {token}'})
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {'users': [user_schema]}
 
 
-def test_update_user(client, user):
+def test_update_user(client, user, token):
     response = client.put(
         '/users/1',
         json={'username': 'bob', 'email': 'bob@example.com', 'password': 'secret'},
+        headers={'Authorization': f'Bearer {token}'}
     )
 
     assert response.status_code == HTTPStatus.OK
@@ -76,13 +70,14 @@ def test_update_user(client, user):
     }
 
 
-def test_update_user_invalid(client):
+def test_update_user_forbidden(client, token):
     response = client.put(
         '/users/2',
         json={'username': 'bob', 'email': 'bob@example.com', 'password': 'secret'},
+        headers={'Authorization': f'Bearer {token}'}
     )
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
+    assert response.status_code == HTTPStatus.FORBIDDEN
 
 
 def test_get_user(client, user):
@@ -98,26 +93,26 @@ def test_get_user_not_found(client):
     assert response.status_code == HTTPStatus.NOT_FOUND
 
 
-def test_delete_user(client, user):
-    response = client.delete('/users/1')
+def test_delete_user(client, user, token):
+    response = client.delete(f'/users/{user.id}', headers={'Authorization': f'Bearer {token}'})
 
     assert response.status_code == HTTPStatus.NO_CONTENT
 
 
-def test_delete_user_invalid(client):
-    response = client.delete('/users/-1')
+def test_delete_user_forbidden(client, user, token):
+    response = client.delete(f'/users/{user.id + 1}', headers={'Authorization': f'Bearer {token}'})
 
-    assert response.status_code == HTTPStatus.NOT_FOUND
+    assert response.status_code == HTTPStatus.FORBIDDEN
 
 
-def test_update_integrity_error(client, user):
+def test_update_integrity_error(client, user, token):
     client.post(
         '/users',
         json={
             'username': 'fausto',
             'email': 'fausto@example.com',
             'password': 'secret',
-        },
+        }
     )
 
     response = client.put(
@@ -127,7 +122,17 @@ def test_update_integrity_error(client, user):
             'email': 'bob@example.com',
             'password': 'mynewpassword',
         },
+        headers={'Authorization': f'Bearer {token}'}
     )
 
     assert response.status_code == HTTPStatus.CONFLICT
     assert response.json() == {'detail': 'Username or Email already exists'}
+
+
+def test_get_token(client, user):
+    response = client.post('/token', data={'username': user.email, 'password': user.clean_password})
+    token = response.json()
+
+    assert response.status_code == HTTPStatus.OK
+    assert token['token_type'] == 'Bearer'
+    assert 'access_token' in token
